{"schema":"decodex/commit/1","summary":"Move cursor icon runtime","authority":"manual"}
